Support loading multiple jqueries
So, the problem here is that jquery is a library that uses define (AMD
style), but passes it an explicit string which is non unique, "jquery".
This means that if you load different jquery versions they will *all*
try to define themselves as "jquery".

This wrapper removes the amd attribute from define, which is used by
jquery to figure out whether to register amd style or not.

See line 10311:
http://code.jquery.com/jquery-1.11.3.js

Yes, this is still a little hacky: other libraries that try to properly
support AMD may check for this var :S