Support loading multiple jqueries
So, the problem here is that jquery is a library that uses define (AMD
style), but passes it an explicit string which is non unique, "jquery".
This means that if you load different jquery versions they will all
try to define themselves as "jquery".

Using the unminified jquery version (we bundle when creating MDAs anyway)
and removing the jquery identifier solves this problem, as the define
will then define itself as the path to the file.